+++ src/app/layout.tsx
@@ -4,6 +4,8 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { twMerge } from "tailwind-merge";
 import { Footer } from "@/components/Footer";
+import { ToggleTheme } from "@/components/ToggleTheme";
+import { ThemeProvider } from "@/components/ThemeProvider";
 
 const inter = Inter({
   subsets: ["latin"],
@@ -21,22 +23,26 @@ export default function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
+  ToggleTheme;
+
   return (
     <html lang="en" className="dark:bg-slate-900">
-      <body
-        className={twMerge(
-          inter.className,
-          "flex antialiased h-screen overflow-hidden bg-gray-100 dark:bg-slate-900"
-        )}
-      >
-        <Sidebar />
-        <div className="lg:pl-2 lg:pt-2 bg-gray-100 dark:bg-slate-900 flex-1 overflow-y-auto">
-          <div className="flex-1 bg-white min-h-screen lg:rounded-tl-xl border border-transparent lg:border-neutral-200 lg:dark:border-slate-800 overflow-y-auto dark:bg-slate-800 -z-5">
-            {children}
-            <Footer />
+      <ThemeProvider>
+        <body
+          className={twMerge(
+            inter.className,
+            "flex antialiased h-screen overflow-hidden bg-gray-100 dark:bg-slate-900"
+          )}
+        >
+          <Sidebar />
+          <div className="lg:pl-2 lg:pt-2 bg-gray-100 dark:bg-slate-900 flex-1 overflow-y-auto">
+            <div className="flex-1 bg-white min-h-screen lg:rounded-tl-xl border border-transparent lg:border-neutral-200 lg:dark:border-slate-800 overflow-y-auto dark:bg-slate-800 -z-5">
+              {children}
+              <Footer />
+            </div>
           </div>
-        </div>
-      </body>
+        </body>
+      </ThemeProvider>
     </html>
   );
 }
